requirements.txt
pillow==9.5.0
tqdm==4.66.1
